opusenc: Consistently use large file interfaces
On 32-bit OS X fseek() fails when seeking to a position >=2^31 (even
fseek(in, 0, SEEK_END)); fseeko() must be used for larger files.
Also, on any platform with a 32-bit long, ftell(in) won't work when
the current file position is >=2^31 (e.g. if there was a very large
skipped chunk).

Address these issues by consistently using the fseek/ftell variants
that can handle large files, when available, for all input file
processing rather than in only one place.  Don't try to seek or use
the result if ftell() returns an error (e.g. EOVERFLOW).  When
reading data until EOF, treat the ftell() value as only an estimate
for percent progress; if the computed size was inaccurate or the
file grows, do not discard data after this point.